+++ holiday-calendar-base/src/test/java/com/github/davejoyce/calendar/TestObjects.java
@@ -18,6 +18,7 @@
 
 package com.github.davejoyce.calendar;
 
+import com.github.davejoyce.calendar.function.DateRoll;
 import com.github.davejoyce.calendar.function.Observance;
 
 import java.time.DayOfWeek;
